Suggestion on net_price member function
Hi, pezy. There seems to be a tiny error in function body of net_price. As mentioned in Exercise 15.7, "If the number of copies exceeds that limit, the normal price applies to those purchased beyond the limit." instead of to all the copies. So I modified the function body of net_price a bit.